Refactor into atomic functions
